Fix for issue #52
When running Doxypypy on a Windows machine, getLines() creates an array of strings separated by '\r\n' while technically correct, when this list is passed to the print() function, is converts the '\n' to a '\r\r\n' sequence causing issues with Doxygen on Windows and text editors in general. By changing the value passed to join() from linesep to '\n', print() properly converts the line feeds to on all platforms.