Do not match string prefix at end of docstring
The original docstring regex removed sequences such as `r"""` from a
docstring regardless if it appeared at the beginning or at the end of
the docstring: `r"""Simulator"""` erroneously becomes `Simulato`.
However, `r` etc. are
[string prefixes](https://docs.python.org/3/reference/lexical_analysis.html#string-and-bytes-literals)
and do not appear at the end of a string.

Therefore, this commit changes the regex to match string prefixes only
at the beginning of the docstring and not at the end.

(Closes #84)